Use ROLLUP operator for fachoursbymonth3 example
The patch also simplifies the where clause to be more readable, but it's
worth noting that the new version will likely prevent postgres from
deriving optimal query plans in the presence of an index on the
starttime column. But IMO performance concerns are beyond the scope of
this project, so I think it's a fair tradeoff.

Note: I did not setup a local dev environment, but hopefully I didn't
break anything ;).

Fixes #14